Removed old logger library and refactored some packages
