+++ pom.xml
@@ -27,7 +27,6 @@
         <jdk.version>1.8</jdk.version>
         <jacoco.version>0.7.6.201602180812</jacoco.version>
         <junit.version>4.12</junit.version>
-        <slf4j.logger.version>1.6.2</slf4j.logger.version>
         <lombok.version>1.16.8</lombok.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
@@ -99,11 +98,6 @@
             <artifactId>junit</artifactId>
             <version>${junit.version}</version>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-simple</artifactId>
-            <version>${slf4j.logger.version}</version>
         </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
