+++ src/main/java/net/liquidpineapple/pang/objects/buttons/SinglePlayerButton.java
@@ -1,6 +1,5 @@
-package net.liquidpineapple.pang.objects;
+package net.liquidpineapple.pang.objects.buttons;
 
-import lombok.extern.slf4j.Slf4j;
 import net.liquidpineapple.pang.Application;
 import net.liquidpineapple.pang.screens.ControlsScreen;
 
@@ -9,7 +8,6 @@ import net.liquidpineapple.pang.screens.ControlsScreen;
  * @author Govert de Gans
  * @date 2016-09-08
  */
-@Slf4j
 public class SinglePlayerButton extends Button {
     /**
      * Constructor of the SinglePlayer button.
