+++ src/test/java/net/liquidpineapple/pang/objects/LifeSystemTest.java
@@ -1,5 +1,6 @@
 package net.liquidpineapple.pang.objects;
 
+import net.liquidpineapple.pang.gui.LifeSystem;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
